requirement.txt
torch
torchvision
tqdm
PyYAML
matplotlib
seaborn
scikit-learn
ipykernel
notebook
torchsummary